Refactor code
Rename tokens-global-raw module into global-raw-tokens
Rename Opacity by OpacityRawTokens
Opacity values are const now
