+++ global-raw-tokens/src/main/AndroidManifest.xml
